requirement.txt
# Core dependencies
easyocr==1.7.2
opencv-python==4.11.0.86
pillow==11.0.0
numpy==2.2.4
matplotlib==3.10.1
torch==2.6.0+cu118
torchvision==0.21.0+cu118
python-bidi==0.6.6
PyYAML==6.0.2
scikit-image==0.25.2
shapely==2.1.0
pyclipper==1.3.0.post6